authors: split blueprints
* Split author module view into holdingpen and publications views.

* Adds basic test to prevent regression.
